requirements.txt
yfinance
pandas
apache-airflow-providers-postgres